working on higher half

+++ src/kernel/clibs/memory.c
@@ -1,7 +1,7 @@
 #include "memory.h"
 #include "stdint.h"
 
-void* memcpy(void* dest,const void* src, uint16_t dest_size){
+void* memcpy(void* dest,const void* src, uint32_t dest_size){
     const uint8_t* copy_src=(const uint8_t*)src;
     uint8_t* copy_dest=(uint8_t*)dest;
     for (int i=0; i<dest_size; i++){
